.ipynb_checkpoints/requirements-checkpoint.txt
streamlit==1.28.1
langchain==0.0.340
langchain-community==0.0.1
chromadb==0.4.15
sentence-transformers==2.2.2
cohere==4.32
google-generativeai==0.3.1
reportlab==4.0.4
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
tiktoken==0.5.1
openai==1.3.5
faiss-cpu==1.7.4
PyPDF2==3.0.1